# repo/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: KuberCloud ANI Core API
  version: 1.0.0
  description: |
    KuberCloud ANI Core — 基础设施平台 REST API 契约（唯一真实来源）。

    ## 文件范围与边界
    本文件（api/openapi/v1.yaml）仅包含 ANI Core 基础设施资源：
    instances（VM/Container/GPU/Sandbox/BM/K8s集群）、volumes/filesystems/objects/
    vector-stores、networks、auth/secrets/encryption、gpu-inventory/dpu-inventory、
    registry/metering/observability/notifications/audit 等。
    ANI Services 资源（模型仓库/推理/知识库/PaaS）在 api/openapi/services/v1.yaml。

    ## URL 前缀规范
    servers[0].url = "https://{host}/api/v1"（前缀在 spec 中定义，Gateway 不做路径变换）。
    版本升级时新建 api/openapi/v2.yaml，本文件路径保持稳定。

    ## SDK 生成（make gen-core-sdk）
    Go: github.com/kubercloud/ani-go | Python: pip install kubercloud-ani
    TypeScript: @kubercloud/ani-client | Java: com.kubercloud:ani-java

    ## 认证
    Bearer JWT（1小时有效）或 X-API-Key Header。
    tenant_id 从 JWT claims 提取，请求体中的 tenant_id 字段将被忽略。

    ## 错误格式
    `{ code: string, message: string, request_id: string, details?: object }`
    标准错误码：UNAUTHORIZED / FORBIDDEN / NOT_FOUND / CONFLICT / BAD_REQUEST /
    RATE_LIMIT_EXCEEDED / NOT_IMPLEMENTED / INTERNAL_ERROR

    ## 分页
    cursor 分页：`?limit=20&cursor=<token>` → `{ items: [], next_cursor: string | null }`

    ## 异步操作
    202 响应体返回 AsyncTask，Location Header 附任务 URL。
    轮询 GET /api/v1/tasks/{task_id} 或配置 Webhook。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: ani.internal
        description: ANI Gateway 地址（支持IP地址）

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT AccessToken（有效期1小时，通过 /auth/token 获取）
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 长期 API Key（在 Console → 设置 → API Key 中创建）

# ── 公共 Schema ───────────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:        { type: string, example: "MODEL_NOT_FOUND" }
        message:     { type: string, example: "模型 qwen2.5-72b 不存在" }
        request_id:  { type: string, example: "req_01j8xabc" }
        details:     { type: object, additionalProperties: true }

    # Cursor 分页（所有列表接口统一使用此格式）
...
    CoreDevProfileInfo:
      type: object
      description: "Core dev/local profile 标记；用于区分本地联调成功与真实 provider 执行成功。"
      required: [mode, provider, real_provider]
      properties:
        mode:          { type: string, enum: [local, real] }
        provider:      { type: string }
        real_provider: { type: boolean }
        reason:        { type: string, nullable: true }
...
    VectorStoreState:
      type: string
      enum: [pending, ready, failed, deleting, deleted]
    VectorStore:
      type: object
      required: [id, tenant_id, name, dimension, metric, state, created_at, updated_at]
      properties:
        id:        { type: string }
        tenant_id: { type: string }
        name:      { type: string }
        dimension: { type: integer, minimum: 1 }
        metric:    { type: string, enum: [cosine, l2, ip] }
        state:     { $ref: '#/components/schemas/VectorStoreState' }
        reason:    { type: string, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    VectorStoreListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/VectorStore' } }
        total: { type: integer }
    CreateVectorStoreRequest:
      type: object
      required: [name, dimension]
      properties:
        name:      { type: string, minLength: 1, maxLength: 128 }
        dimension: { type: integer, minimum: 1 }
        metric:    { type: string, enum: [cosine, l2, ip], default: cosine }
    VectorStoreSearchRequest:
      type: object
      required: [vector]
      properties:
        vector:
          type: array
          minItems: 1
          items: { type: number, format: float }
        top_k: { type: integer, minimum: 1, maximum: 100, default: 10 }
        filter:
          type: object
          additionalProperties: { type: string }
    VectorStoreSearchHit:
      type: object
      required: [id, score, metadata]
      properties:
        id:       { type: string }
        score:    { type: number, format: float }
        metadata:
          type: object
          additionalProperties: { type: string }
    VectorStoreSearchResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/VectorStoreSearchHit' } }
        total: { type: integer }
...
  responses:
    Unauthorized:
      description: 未认证或 Token 无效（code=UNAUTHORIZED）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 无权访问该资源（code=FORBIDDEN）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在（code=NOT_FOUND）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: 请求参数不合法（code=BAD_REQUEST）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── API 路径定义 ──────────────────────────────────────────────────────────────
# 说明：/demo/* 路径已迁移至 api/openapi/dev.yaml（不参与 SDK 生成）
paths:
...
  /vector-stores:
    get:
      operationId: listVectorStores
      summary: 查询向量存储列表
      tags: [VectorStore]
      x-ani-rbac-scope: "scope:vector-stores:read"
      responses:
        "200":
          description: 向量存储列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VectorStoreListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createVectorStore
      summary: 创建向量存储
      tags: [VectorStore]
      x-ani-rbac-scope: "scope:vector-stores:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVectorStoreRequest' }
      responses:
        "201":
          description: 向量存储已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VectorStore' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /vector-stores/{vector_store_id}:
    get:
      operationId: getVectorStore
      summary: 查询向量存储
      tags: [VectorStore]
      x-ani-rbac-scope: "scope:vector-stores:read"
      parameters:
        - { name: vector_store_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 向量存储
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VectorStore' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteVectorStore
      summary: 删除向量存储
      tags: [VectorStore]
      x-ani-rbac-scope: "scope:vector-stores:delete"
      parameters:
        - { name: vector_store_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 向量存储已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VectorStore' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /vector-stores/{vector_store_id}/search:
    post:
      operationId: searchVectorStore
      summary: 搜索向量存储
      tags: [VectorStore]
      x-ani-rbac-scope: "scope:vector-stores:search"
      parameters:
        - { name: vector_store_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VectorStoreSearchRequest' }
      responses:
        "200":
          description: 向量检索结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VectorStoreSearchResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── 平台品牌化 ──────────────────────────────────────────────────────────────
...